Move get_cluster_info() and get_fixed_params() to utils as it can be used in the scripts.
