+++ ga_dashboard/backend/utils.py
@@ -1,8 +1,12 @@
+import argparse
 import datetime
+import os
 import sys
 import random
 import pandas as pd
 import numpy as np
+import yaml
+
 
 class validate_args:
     """
@@ -146,3 +150,32 @@ def simulate_mock_jobs(): # DEBUGONLY
         df_list.append(data_frame)
 
     return pd.concat(df_list)
+
+def get_cluster_info(ns: argparse.Namespace, info_file: str) -> object:
+    """
+    Get the YAML object representation of a cluster info file.
+    :param ns: [argparse.Namespace] Namespace representing the command-line arguments.
+    :param info_file: [str] Name of info file, e.g., 'cluster_info.yaml'
+    :return: [argparse.Namespace] The populated Namespace (args) object.
+    """
+    with open(os.path.join(ns.path_infrastructure_info, info_file), "r") as stream:
+        try:
+            cluster_info = yaml.safe_load(stream)
+        except yaml.YAMLError as exc:
+            print(exc)
+    return cluster_info
+
+
+def get_fixed_params(ns: argparse.Namespace, fp_file: str) -> object:
+    """
+    Get the YAML object representation of the fixed parameters file.
+    :param ns: [argparse.Namespace] Namespace representing the command-line arguments.
+    :param fp_file: [str] Name of fixed params file, e.g., 'cluster_info.yaml'
+    :return: [argparse.Namespace] The populated Namespace (args) object.
+    """
+    with open(os.path.join(ns.path_infrastructure_info, fp_file), "r") as stream:
+        try:
+            fixed_params = yaml.safe_load(stream)
+        except yaml.YAMLError as exc:
+            print(exc)
+    return fixed_params
